move methods to Function interface
