+++ pulsar-functions/instance/src/main/java/org/apache/pulsar/functions/instance/JavaInstance.java
@@ -32,7 +32,6 @@ import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.pulsar.functions.api.Function;
 import org.apache.pulsar.functions.api.Record;
-import org.apache.pulsar.functions.api.InitializableFunction;
 
 /**
  * This is the Java Instance. This is started by the runtimeSpawner using the JavaInstanceClient
@@ -157,11 +156,8 @@ public class JavaInstance implements AutoCloseable {
     }
 
     public void initialize() throws Exception {
-        if (function instanceof InitializableFunction) {
-            ((InitializableFunction) function).initialize(context);
-        }
-        if (javaUtilFunction instanceof InitializableFunction) {
-            ((InitializableFunction) javaUtilFunction).initialize(context);
+        if (function != null) {
+            function.initialize(context);
         }
     }
 
@@ -170,16 +166,9 @@ public class JavaInstance implements AutoCloseable {
         context.close();
         executor.shutdown();
 
-        if (function instanceof InitializableFunction) {
+        if (function != null) {
             try {
-                ((InitializableFunction) function).close();
-            } catch (Exception e) {
-                log.error("function closeResource occurred exception", e);
-            }
-        }
-        if (javaUtilFunction instanceof InitializableFunction) {
-            try {
-                ((InitializableFunction) javaUtilFunction).close();
+                function.close();
             } catch (Exception e) {
                 log.error("function closeResource occurred exception", e);
             }
