--- pulsar-functions/java-examples/src/main/java/org/apache/pulsar/functions/api/examples/InitFunction.java
@@ -1,47 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.pulsar.functions.api.examples;
-
-import lombok.extern.slf4j.Slf4j;
-import org.apache.pulsar.functions.api.Context;
-import org.apache.pulsar.functions.api.InitializableFunction;
-
-@Slf4j
-public class InitFunction implements InitializableFunction<String, String> {
-
-    public static boolean initialized = false;
-
-    @Override
-    public void initialize(Context context) throws Exception {
-        initialized = true;
-    }
-
-    @Override
-    public void close() throws Exception {
-        initialized = false;
-    }
-
-    @Override
-    public String process(String input, Context context) throws Exception {
-        if (!initialized) {
-            throw new Exception("function not initialized");
-        }
-        return "message";
-    }
-}
